+++ install.py
@@ -1,7 +1,10 @@
-
-# installer derived from the weewx Belchertown skin installer
-# https://raw.githubusercontent.com/poblabs/weewx-belchertown/master/install.py
-# which was Copyright Pat O'Brien, with re-fomatting from a PR by Vince Skahan 
+"""
+Credits:
+ Installer derived from https://github.com/smeisens/weewx-wundergroundlike
+ in turn derived from the weewx Belchertown skin installer
+ https://raw.githubusercontent.com/poblabs/weewx-belchertown/master/install.py
+ which was Copyright Pat O'Brien, with re-fomatting from a PR by Vince Skahan 
+"""
 
 import configobj
 from setup import ExtensionInstaller
@@ -11,20 +14,20 @@ from io import StringIO
 
 #-------- extension info -----------
 
-VERSION      = "1.0.0"
-NAME         = 'wundergroundLike'
-DESCRIPTION  = 'Post to a custom server_url that uses Weather Underground formatting'
-AUTHOR       = "Sigi Meisenbichler"
-AUTHOR_EMAIL = "s.meisen@icloud.com"
+VERSION      = "1.0.1"
+NAME         = 'OgoxeUploader'
+DESCRIPTION  = 'Post to the OgoXe weather platform'
+AUTHOR       = "OgoXe SAS"
+AUTHOR_EMAIL = "dev@ogoxe.com"
 
 #-------- main loader -----------
 
 def loader():
-    return WundergroundLikeInstaller()
+    return OgoxeUploaderInstaller()
 
-class WundergroundLikeInstaller(ExtensionInstaller):
+class OgoxeUploaderInstaller(ExtensionInstaller):
     def __init__(self):
-        super(WundergroundLikeInstaller, self).__init__(
+        super(OgoxeUploaderInstaller, self).__init__(
             version=VERSION,
             name=NAME,
             description=DESCRIPTION,
@@ -43,23 +46,21 @@ extension_config = """
 
 [StdRESTful]
 
-    [[WundergroundLike]]
-        # This section is for configuring posts to a site that expects
-        # Weather Underground formatted data. This typically is only
-        # needed if you post to the actual Weather Underground 'and'
-        # you 'also' want to post similarly to a different site.
+    [[OgoxeUploader]]
+        # This section is for configuring posts to the Ogoxe Weather Platform
+        # that expects Weather Underground formatted data (without the "Rapidfire" protocol). 
 
-        # Note - the WU "Rapidfire" protocol is 'disabled' in this extension
-        # but other optional parameters supported by weewx-5.2.0 'should' work.
+        # The current uploader relies on WeeWX's builtin Weather Underground uploader,
+        # thus, optional parameters supported by weewx-5.2.0 'should' work.
         # Please consult the weewx documentation for details.
 
         # To enable this uploader, set enable = True below and specify your
-        # appropriate values for station, password, and server_url
+        # appropriate values for station, password.
+        # The server url is preset and cannot be modified.
 
         enable = false
         station = replace_me
         password = replace_me
-        server_url = replace_me
 
 """
 config_dict = configobj.ConfigObj(StringIO(extension_config))
@@ -67,10 +68,10 @@ config_dict = configobj.ConfigObj(StringIO(extension_config))
 #----------------------------------
 #  files and services stanzas
 #----------------------------------
-files=[('bin/user', ['bin/user/wundergroundLike.py'])]
+files=[('bin/user', ['bin/user/ogoxeUploader.py'])]
 files_dict = files
 
-restful_services = ['user.wundergroundLike.WundergroundLike']
+restful_services = ['user.ogoxeUploader.OgoxeUploader']
 restful_dict = restful_services
 
 #---------------------------------